requirements.txt
newspaper3k==0.2.8
selenium
